settings: refactor system resource panel
