+++ ui/src/state/azimuth.ts
@@ -1,44 +1,48 @@
-import { useCallback, useEffect, useState } from 'react';
+import { useEffect } from 'react';
+import { useAsyncCall } from '../logic/useAsyncCall';
 import { fakeRequest } from './util';
 
 const api = {
   forceUpdate() {
     return fakeRequest(null, 1000);
   },
-  async getAzimuthBlock() {
-    return fakeRequest('16.514.728', 1000);
+  async getAzimuthState() {
+    return fakeRequest(
+      {
+        block: '16.514.728',
+        stale: false,
+      },
+      1000
+    );
   },
 };
 
-export const useAzimuthBlock = () => {
-  const [isLoading, setIsLoading] = useState(false);
-  const [isStale, setIsStale] = useState(true);
-  const [block, setBlock] = useState<string | null>(null);
+export const useAzimuthState = () => {
+  const {
+    status: stateLoadStatus,
+    call: loadState,
+    error: stateLoadError,
+    result: state,
+  } = useAsyncCall(api.getAzimuthState);
 
-  const loadBlock = useCallback(async () => {
-    setIsLoading(true);
-    const block = await api.getAzimuthBlock();
-    setBlock(block);
-    setIsLoading(false);
-  }, []);
+  const {
+    status: forceUpdateStatus,
+    call: forceUpdate,
+    error: forceUpdateError,
+  } = useAsyncCall(api.forceUpdate);
 
-  const forceUpdate = useCallback(async () => {
-    setIsLoading(true);
-    setIsStale(false);
-    await api.forceUpdate();
-    setIsLoading(false);
-  }, []);
-
+  // attempt to load initial State on mount
   useEffect(() => {
-    // attempt to load initial block on mount
-    loadBlock();
+    loadState();
   }, []);
 
   return {
-    isLoading,
-    isStale,
-    block,
+    state,
+    stateLoadStatus,
+    stateLoadError,
     forceUpdate,
+    forceUpdateStatus,
+    forceUpdateError,
   };
 };
 [];
