+++ ui/src/state/connectivity.ts
@@ -1,10 +1,13 @@
-import { useCallback, useEffect, useState } from 'react';
+import { useEffect } from 'react';
+import { useAsyncCall } from '../logic/useAsyncCall';
 import { fakeRequest } from './util';
 
 export type AvailabilityStatus = 'initial' | 'available' | 'unavailable';
 
 const api = {
-  checkShipAvailability(shipName: string) {
+  checkShipAvailability(
+    shipName: string
+  ): Promise<{ status: AvailabilityStatus }> {
     return fakeRequest({
       status: 'available',
     });
@@ -14,49 +17,32 @@ const api = {
   },
 };
 
-const checkShipAvailability = async (
-  shipName: string
-): Promise<{ status: AvailabilityStatus }> => {
-  return fakeRequest({
-    status: 'available',
-  });
-};
-
 export const useShipAvailability = (shipName: string) => {
-  const [availabilityStatus, setAvailabilityStatus] =
-    useState<AvailabilityStatus>('initial');
-  const [isChecking, setIsChecking] = useState(false);
-
-  const checkAvailability = useCallback(async () => {
-    setIsChecking(true);
-    const { status } = await checkShipAvailability(shipName);
-    setAvailabilityStatus(status);
-    setIsChecking(false);
-  }, []);
+  const {
+    call: checkAvailability,
+    result: availability,
+    status,
+    error,
+  } = useAsyncCall(api.checkShipAvailability);
 
   useEffect(() => {
-    checkAvailability();
+    checkAvailability(shipName);
   }, []);
 
-  return { isChecking, availabilityStatus, refresh: checkAvailability };
+  return { status, availability, error, checkAvailability };
 };
 
 export const usePeerDiscoveryShips = () => {
-  const [isLoading, setIsLoading] = useState(false);
-  const [peerDiscoveryShips, setPeerDiscoverShips] = useState<string[] | null>(
-    null
-  );
-
-  const getShips = useCallback(async () => {
-    setIsLoading(true);
-    const ships = await api.getPeerDiscoveryShips();
-    setPeerDiscoverShips(ships);
-    setIsLoading(false);
-  }, []);
+  const {
+    call: getShips,
+    result: peerDiscoveryShips,
+    status,
+    error,
+  } = useAsyncCall(api.getPeerDiscoveryShips);
 
   useEffect(() => {
     getShips();
   }, []);
 
-  return { isLoading, peerDiscoveryShips, getShips };
+  return { status, error, peerDiscoveryShips, getShips };
 };
